# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wossolaty API Documentation",
        "description": "This is a wossolaty API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "haidershamil@gmail"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment_types": {
            "get": {
                "tags": [
                    "Payment_Type"
                ],
                "summary": "Returns API response of all Payment_Type",
                "description": "A sample test of the API",
                "operationId": "getPayment_TypeList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment_Type"
                ],
                "summary": "Store new Payment_Type",
                "description": "Returns Payment_Type data",
                "operationId": "storePayment_Type",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/payment_types/{id}": {
            "get": {
                "tags": [
                    "Payment_Type"
                ],
                "summary": "Get Payment_Type information",
                "description": "Returns Payment_Type data",
                "operationId": "getPayment_TypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payment_Type"
                ],
                "summary": "Update existing Payment_Type",
                "description": "Returns updated Payment_Type data",
                "operationId": "updatePayment_Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payment_Type"
                ],
                "summary": "Delete existing Payment_Type",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePayment_Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
}